BJS2-27337 bug fixes

+++ src/test/java/faang/school/accountservice/controller/SavingsAccountControllerTest.java
@@ -13,6 +13,7 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import static org.hamcrest.Matchers.hasSize;
@@ -44,7 +45,7 @@ class SavingsAccountControllerTest {
         savingsAccountId = 1L;
         savingsAccountDto = new SavingsAccountDto();
         savingsAccountDto.setId(savingsAccountId);
-        savingsAccountDto.setRate(1.8);
+        savingsAccountDto.setRate(BigDecimal.valueOf(1.8));
         savingsAccountDto.setTariffId(3L);
     }
